code refactor - added seperate folder for services and moved patientPoratlService calls to new folder
