+++ src/client/components/patientPortal/dashboard/profile/index.jsx
@@ -1,7 +1,7 @@
 import React, { Component } from "react";
 import DashboardSidebar from "../sidebar/sidebar.jsx";
-//import IMG01 from '../../../../assets/images/patients/patient.jpg';
 import StickyBox from "react-sticky-box";
+
 class Profile extends Component {
 	render() {
 		return (
