+++ src/client/patientPortalServices/constants.js
@@ -1,12 +1,7 @@
 export const serviceConstants = {
-    // HOST_NAME: 'http://3.137.173.35:6061'
-    HOST_NAME: 'https://www.mycovidnow.com',
-   API_HOST_NAME: 'https://www.mycovidnow.com/api',
-
-    //  HOST_NAME: 'http://localhost:8085',
-    //  API_HOST_NAME: 'http://localhost:8085',
-
-    GOOGLE_API_KEY : "AIzaSyCq8vVmt0ls7bTWUoqBhbITd5_Xvi8u2qY",
-
-    AG_GRID_LICENSE_KEY : "CompanyName=American Gene Technologies,LicensedApplication=MyCovidNow_AGT,LicenseType=SingleApplication,LicensedConcurrentDeveloperCount=1,LicensedProductionInstancesCount=0,AssetReference=AG-014604,ExpiryDate=28_March_2022_[v2]_MTY0ODQyMjAwMDAwMA==6a5adf639c7dd8b8b7dde3799857052d"
+	HOST_NAME: "https://www.mycovidnow.com",
+	API_HOST_NAME: "https://www.mycovidnow.com/api",
+	GOOGLE_API_KEY: "AIzaSyCq8vVmt0ls7bTWUoqBhbITd5_Xvi8u2qY",
+	AG_GRID_LICENSE_KEY:
+		"CompanyName=American Gene Technologies,LicensedApplication=MyCovidNow_AGT,LicenseType=SingleApplication,LicensedConcurrentDeveloperCount=1,LicensedProductionInstancesCount=0,AssetReference=AG-014604,ExpiryDate=28_March_2022_[v2]_MTY0ODQyMjAwMDAwMA==6a5adf639c7dd8b8b7dde3799857052d",
 };
